updated local variables

+++ src/main/java/com/auth_service/config/properties/JwtConfigurationProperties.java
@@ -3,9 +3,9 @@ package com.auth_service.config.properties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @ConfigurationProperties(prefix = "jwt")
-public record JwtConfigurationProperties(
-        String secret,
-        Long accessExpirationTime,
-        Long refreshExpirationTime
-        ) {
-}
+public record JwtConfigurationProperties
+        (
+                String secret,
+                Integer accessExpirationTime,
+                Integer refreshExpirationTime
+        ) {}
